VB-1169 Make bft-engine replica working with real persistency code. (#130)
Integrate persistency code with a file-based storage (no restarts yet).

* Fix BlsMultisigVerifier deserialization.
* Fix replica load from file.
* Fix creation of Signer/Verifier classes.
* Add more strict checks for simple parameters.
* Continue replacing raw pointers by smart.
* Fix support for re-entrant writes.
* Fix sequence number size.
* Make simpleTest use a file storage for metadata parameters.
* Make FileStorage ignore objectId=0, which is used as DB initialization flag.
* Fix product compilation issues + increase possible number of metadata objects.

+++ bftengine/tests/simpleStorage/ObjectsMetadataHandler.cpp
@@ -36,11 +36,11 @@ size_t ObjectsMetadataHandler::getObjectsMetadataSize() {
 
 void ObjectsMetadataHandler::setObjectInfo(
     const MetadataObjectInfo &objectInfo) {
-  objectsMap_.insert(pair<uint16_t, MetadataObjectInfo>(objectInfo.id,
+  objectsMap_.insert(pair<uint32_t, MetadataObjectInfo>(objectInfo.id,
                                                         objectInfo));
 }
 
-MetadataObjectInfo *ObjectsMetadataHandler::getObjectInfo(uint16_t objectId) {
+MetadataObjectInfo *ObjectsMetadataHandler::getObjectInfo(uint32_t objectId) {
   auto it = objectsMap_.find(objectId);
   if (it != objectsMap_.end()) {
     return &(it->second);
